Separate logger to util; avoiding circ dependencies
