+++ data-serving/data-service/src/index.ts
@@ -27,20 +27,9 @@ import dotenv from 'dotenv';
 import express from 'express';
 import expressStatusMonitor from 'express-status-monitor';
 import mongoose from 'mongoose';
-import pino from 'pino';
 import swaggerUi from 'swagger-ui-express';
 import validateEnv from './util/validate-env';
-
-// The Pino team recommends against using pino-pretty/prettyPrint in
-// production. Without a prettifier, Pino logs JSON messages. This is faster,
-// and decouples log writing and reading concerns, but requires that viewers
-// install pino-pretty globally and pipe all logs commands to it.
-//
-// Instead, prettifying programmatically, since we only use it for HTTP errors,
-// as opposed to all requests.
-export const logger = pino({
-    prettyPrint: { colorize: process.env.NODE_ENV !== 'production' },
-});
+import { logger } from './util/logger';
 
 const app = express();
